+++ src/App.jsx
@@ -2,7 +2,7 @@ import "./App.css";
 import React, { useState, useEffect } from "react";
 import { supabase } from "./createClient";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
-import Home from "./Pages/Home";
+import Home from "./Pages/Home/Home.jsx";
 import Location from "./Pages/Location";
 import Errorpage from "./Pages/Errorpage.jsx";
 
@@ -10,7 +10,7 @@ function App() {
 
   return (
 
-    <>
+  <>
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Home />} />
